% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ocmm.methods.R
\name{anova.ocmm}
\alias{anova.ocmm}
\title{Anova method for Continuous Ordinal Fits}
\usage{
\method{anova}{ocmm}(object, ...)
}
\arguments{
\item{object}{An ocm object.}

\item{...}{one or more additional ocm objects.}
}
\value{
The method returns an object of class \code{anova.ocmm} and \code{data.frame}, reporting for each model, in hierarchical order:
\itemize{
  \item no.par the number of parameters
  \item AIC the Akaike information criterion
  \item loglik the log-likelihood
  \item LR.stat the likelihood ratio statistic
  \item df the difference in the degrees of freedom in the models being compared
  \item Pr(>Chisq) the p-value from the likelihood ratio test
}
}
\description{
Comparison of continuous ordinal models in likelihood ratio tests.
}
\examples{
\dontrun{
fit.overall.rnd  <- ocmm(overall  ~ cycleno + bsa + treatment + (1|randno), data=ANZ0001)
anova(fit.overall.rnd, update(fit.overall, .~. + age))
}
}
\keyword{anova}

