% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ocm.methods.R
\name{coef.ocm}
\alias{coef.ocm}
\title{Extract Model Coefficients}
\usage{
coef(object, ...)
}
\arguments{
\item{object}{an ocm object.}

\item{...}{other arguments.}
}
\description{
Extract Model Coefficients
}

