% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ocm_gen.R
\name{deriv_link}
\alias{deriv_link}
\title{Function to compute the derivatives of the link function needed by the algorithm}
\usage{
deriv_link(link = c("logit", "probit", "cloglog", "loglog", "cauchit"))
}
\arguments{
\item{link}{One of "logit" (default), "probit", "cloglog", "loglog" or "cauchit".}
}
\value{
A list with the link function and the 1st, 2nd and 3rd derivatives with respect to the argument
}
\description{
Function to compute the derivatives of the link function needed by the algorithm
}
