% Generated by roxygen2 (4.1.0.9000): do not edit by hand
% Please edit documentation in R/ocm.methods.R
\name{extractAIC.ocm}
\alias{extractAIC.ocm}
\title{Extract AIC from a Fitted Model}
\usage{
\method{extractAIC}{ocm}(fit, scale = 0, k = 2, ...)
}
\arguments{
\item{fit}{fitted model.}

\item{scale}{parameter currently not used. For compatibility with general extractAIC method.}

\item{k}{numeric specifying the ‘weight’ of the equivalent degrees of freedom (=: edf) part in the AIC formula. Defaults to 2.}

\item{...}{further arguments (currently unused)}
}
\description{
Extract AIC from a Fitted Model
}

