% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ocm.methods.R
\name{fitted.ocm}
\alias{fitted.ocm}
\title{Extract Model Fitted Values}
\usage{
\method{fitted}{ocm}(object, ...)
}
\arguments{
\item{object}{an object of class \code{ocm}, usually, a result of a call to \code{ocm}.}

\item{...}{further arguments passed to or from other methods.}
}
\value{
Fitted values computed from \code{object}.
}
\description{
\code{fitted.ocm} is the ordinalCont specific method for the generic function \code{fitted}, 
which computes model fitted from objects of class \code{ocm}.
}
\details{
An object of class \code{ocm} is used to compute the probability 
densities of the continuous ordinal score. The fitted values are the means of such
probability density functions. The output is scaled following the original scale of the scores.
}
\author{
Maurizio Manuguerra, Gillian Heller
}
