% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ocm.methods.R
\name{formula.ocm}
\alias{formula.ocm}
\title{Model Formulae}
\usage{
\method{formula}{ocm}(x, ...)
}
\arguments{
\item{x}{an object of class \code{ocm}, usually, a result of a call to \code{ocm}.}

\item{...}{further arguments passed to or from other methods.}
}
\value{
A symbolic model formula extracted from the model object.
}
\description{
\code{formula.ocm} is the ordinalCont specific method for the generic function \code{formula}, 
which extracts the model formula from objects of class \code{ocm}.
}
\author{
Maurizio Manuguerra, Gillian Heller
}
