% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ocm_gen.R
\name{inv_link}
\alias{inv_link}
\title{Function to compute inverse link functions}
\usage{
inv_link(link = c("logit", "probit", "cloglog", "loglog", "cauchit"))
}
\arguments{
\item{link}{One of "logit" (default), "probit", "cloglog", "loglog" or "cauchit".}
}
\value{
A list with the link function and the 1st, 2nd and 3rd derivatives with respect to the argument
}
\description{
Function to compute inverse link functions
}
