% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ocm.methods.R
\name{logLik.ocm}
\alias{logLik.ocm}
\title{Extract the Log-Likelihood}
\usage{
\\method{logLik}{ocm}(object, ...)
}
\arguments{
\item{object}{an \code{ocm} object.}

\item{...}{Further arguments to be passed to methods.}
}
\value{
Returns the log-likelihood of an \code{ocm} object
}
\description{
Extract the Log-Likelihood
}
\seealso{
\code{\link{ocm}}
}

