% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ocmm.methods.R
\name{logLik.ocmm}
\alias{logLik.ocmm}
\title{Extract the Log-Likelihood}
\usage{
\method{logLik}{ocmm}(object, ...)
}
\arguments{
\item{object}{an \code{ocmm} object.}

\item{...}{Further arguments to be passed to methods.}
}
\value{
Returns the log-likelihood of an \code{ocmm} object
}
\description{
Extract the Log-Likelihood
}
\examples{
\dontrun{
fit.overall.rnd  <- ocmm(overall  ~ cycleno + age + bsa + treatment + (1|randno), data=ANZ0001)
logLik(fit.overall.rnd)
}
}
\seealso{
\code{\link{ocmm}}
}

