% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ocm.methods.R
\name{model.frame.ocm}
\alias{model.frame.ocm}
\title{Model Frame}
\usage{
\method{model.frame}{ocm}(formula, data, random.terms = TRUE, ...)
}
\arguments{
\item{formula}{a model formula}

\item{data}{a data.frame containing the variables in formula.}

\item{random.terms}{a logical indicating if random terms have to be included in the terms object. Defaults to TRUE.}

\item{...}{a mix of further arguments to pass to the default method.}
}
\value{
A c("data.frame") with the variables needed to obtain \code{object}.
}
\description{
\code{model.frame.ocm} is the ordinalCont specific method for the generic function \code{model.frame}, 
which return a \link{data.frame} with the variables needed to use \code{formula} and any ... arguments.
}
\author{
Maurizio Manuguerra, Gillian Heller
}
