% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ocm.methods.R
\name{model.matrix.ocm}
\alias{model.matrix.ocm}
\title{Model Matrix}
\usage{
\method{model.matrix}{ocm}(object, random.terms = TRUE, ...)
}
\arguments{
\item{object}{an object of class \code{ocm}, usually, a result of a call to \code{ocm}.}

\item{random.terms}{a logical indicating if random terms have to be included in the terms object. Defaults to TRUE.}

\item{...}{further arguments passed to or from other methods.}
}
\value{
A design (or model) matrix with the variables needed to obtain the object \code{x}, e.g., by expanding factors to a set of dummy variables and expanding interactions similarly.
}
\description{
\code{model.matrix.ocm} is the ordinalCont specific method for the generic function \code{model.matrix}, 
which extracts the model matrix from objects of class \code{ocm}.
}
\author{
Maurizio Manuguerra, Gillian Heller
}
