% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ocm.R
\name{negloglik4}
\alias{negloglik4}
\title{Penalized log-likelihood function}
\usage{
negloglik4(par, v, pars_obj, wts)
}
\arguments{
\item{par}{vector of regression coefficients}

\item{v}{vector of standardized scores from the continuous ordinal scale}

\item{pars_obj}{the current object of class \code{ocmpars}.}

\item{wts}{optional case weights}
}
\value{
Minus the penalized log-likelihood at parameter values \code{par}
}
\description{
Computes the penalized log-likelihood function
}
\details{
This function computes minus the penalized log-likelihood function. It is used internally 
to fit the model and should not be of interest of the user.
}
\author{
Maurizio Manuguerra, Gillian Heller
}
\keyword{likelihood,}
\keyword{log-likelihood.}
