% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ocm.methods.R
\name{nobs.ocm}
\alias{nobs.ocm}
\title{Extract Model Coefficients}
\usage{
\method{nobs}{ocm}(object, ...)
}
\arguments{
\item{object}{an object of class \code{ocm}, usually, a result of a call to \code{ocm}.}

\item{...}{further arguments passed to or from other methods.}
}
\value{
The (numeric) number of observations in the model object.
}
\description{
\code{nobs.ocm} is the ordinalCont specific method for the generic function \code{nobs}, 
which returns number of observations from objects of class \code{ocm}.
}
\author{
Maurizio Manuguerra, Gillian Heller
}
