% Generated by roxygen2 (4.1.0.9000): do not edit by hand
% Please edit documentation in R/functions.R
\name{ocm.formula}
\alias{ocm.formula}
\title{Continuous ordinal regression}
\usage{
\method{ocm}{formula}(formula, data = list(), ...)
}
\arguments{
\item{formula}{A formula object (fixed effects).}
}
\description{
This function performs the comtinuous ordinal regression with logt link using the generalized logistic function as g function and without random effects.
}
\examples{
ocm.formula()
}
\keyword{likelihood,}
\keyword{log-likelihood.}

