% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ordinalCont.R
\docType{data}
\name{pain}
\alias{pain}
\title{Pain data}
\format{A data frame with 252 rows and 13 variables}
\usage{
data(pain)
}
\description{
Pain data
}
\details{
A dataset containing the VAS scores in a clinical trial for the effectiveness
of laser treatment for chronic neck pain. Patients (n=84) were randomly assigned to receive
either active laser treatment or placebo. VAS scores were recorded by the patients
 at three time points: baseline,
the end of the course of treatment and one month after completion of the treatment.

The variables are as follows:
\tabular{rlll}{
 \tab \bold{Name}\tab \bold{Type}\tab \bold{Description}\cr
   \tab time\tab numeric\tab the time mark at which the VAS score is recorded\cr
  \tab ID\tab numeric\tab The ID number of the patient\cr
  \tab sex\tab ??\tab Sex of patient\cr
  \tab age\tab numeric\tab Age of patient at entry in the study\cr
  \tab group\tab  ?? \tab Trial group (2002 or 2003)\cr
  \tab duration\tab numeric \tab duration of ...\cr
  \tab laterali \tab binary\tab uni- or bi-lateral pain\cr
  \tab laser \tab binary\tab 1 if the laser was active, 0 otherwise\cr
  \tab localisa\tab ??\tab diffuse or localized pain\cr
  \tab vas\tab numeric\tab VAS score (0--1)\cr
  \tab lasert1\tab binary\tab 1 if the laser was active and the VAS score was recorded at baseline, 0 otherwise\cr
  \tab lasert2\tab binary\tab 1 if the laser was active and the VAS score was recorded at the end of the course of treatment, 0 otherwise\cr
  \tab lasert3\tab binary\tab 1 if the laser was active and the VAS score was recorded one month after completion of the treatment, 0 otherwise\cr
 }
}
\references{
Chow RT, Heller GZ, Barnsley L (2006). The effect of 300mW, 830nm laser on
 chronic neck pain: A double-blind, randomized, placebo-controlled study, \emph{Pain},
 124: 201-210.
}
\keyword{datasets}

