% Generated by roxygen2 (4.1.0.9000): do not edit by hand
% Please edit documentation in R/ocmm.methods.R
\name{plot.ocmm}
\alias{plot.ocmm}
\title{Plot method for Continuous Ordinal Fits}
\usage{
\method{plot}{ocmm}(x, CIs = c("no", "vcov"), R = 1000,
  main = "g function (95\% CIs)", xlab = "Continuous ordinal scale",
  ylab = "", ...)
}
\arguments{
\item{x}{An ocm object.}

\item{CIs}{Indicates if confidence bands for the g function should be computed (based on the Wald 95\% CIs).}

\item{R}{The number of bootstrap replicates.}

\item{...}{Further arguments passed to or from other methods.}
}
\description{
This function plots the g function as fitted in an ocm call.
}
\keyword{plot}

