% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ocm.methods.R
\name{plot.predict.ocm}
\alias{plot.predict.ocm}
\title{Plot the probability densities as from the output of the predict method}
\usage{
\method{plot}{predict.ocm}(x, records = NULL, ...)
}
\arguments{
\item{x}{An object of class "predict.ocm".}

\item{records}{A integer or a vector of integers. The number of the record/s
in the data set for which the density has t be plotted. If not specified, the
function will iteratively plot all of them.}

\item{...}{Further arguments passed to or from other methods.}
}
\description{
plot method for class "predict.ocm"
}
\keyword{plot}
\keyword{predict,}

