% Generated by roxygen2 (4.1.0.9000): do not edit by hand
% Please edit documentation in R/ocmm.methods.R
\name{plot.predict.ocmm}
\alias{plot.predict.ocmm}
\title{Plot the probability densities as from the output of the predict method}
\usage{
\method{plot}{predict.ocmm}(x, ...)
}
\arguments{
\item{x}{An object of class "predict.ocm".}

\item{...}{Further arguments passed to or from other methods.}
}
\description{
plot method for class "predict.ocm"
}
\keyword{plot}
\keyword{predict,}

