% Generated by roxygen2 (4.1.0.9000): do not edit by hand
% Please edit documentation in R/functions.R
\name{predict.contOrd}
\alias{predict.contOrd}
\title{Predict method for Continuous Ordinal Fits}
\usage{
\method{predict}{contOrd}(object, newdata = NULL, ...)
}
\arguments{
\item{love}{Do you love cats? Defaults to TRUE.}
}
\description{
Predicted values based on contOrd object.
}
\examples{
cat_function()
}
\keyword{cats}

