% Generated by roxygen2 (4.1.0.9000): do not edit by hand
% Please edit documentation in R/ocmm.methods.R
\name{predict.ocmm}
\alias{predict.ocmm}
\title{Predict method for Continuous Ordinal Fits}
\usage{
\method{predict}{ocmm}(object, newdata = NULL, ...)
}
\arguments{
\item{object}{An ocm object.}

\item{newdata}{optionally, a data frame in which to look for variables with which to predict. Note that all predictor variables should be present having the same names as the variables used to fit the model.}

\item{...}{Further arguments passed to or from other methods.}
}
\description{
Predicted values based on ocm object. [Gillian: we put the rnd effeccts to zero? Should we include this method at all?]
}
\keyword{predict}

