% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ocm.methods.R
\name{print.anova.ocm}
\alias{print.anova.ocm}
\title{Print anova.ocm objects}
\usage{
\method{print}{anova.ocm}(x, digits = max(getOption("digits") - 2, 3),
  signif.stars = getOption("show.signif.stars"), ...)
}
\arguments{
\item{x}{an object of class \code{anova.ocm}}

\item{digits}{controls the number of digits to print. Defaults to the maximum of the value 
returned by (getOption("digits") - 2) and 3}

\item{signif.stars}{a logical. Should the significance stars be printed? Defaults to the value 
returned by getOption("show.signif.stars")}

\item{...}{further arguments passed to or from other methods}
}
\value{
Prints \code{anova.ocm} object
}
\description{
Print the results of the comparison of continuous ordinal models in likelihood ratio tests.
}
\seealso{
\code{\link{ocm}}, \code{\link{anova.ocm}}
}
\author{
Maurizio Manuguerra, Gillian Heller
}
\keyword{anova}
\keyword{summary,}
