% Generated by roxygen2 (4.1.0.9000): do not edit by hand
% Please edit documentation in R/functions.R
\name{print.contOrd}
\alias{print.contOrd}
\title{Continuous ordinal regression}
\usage{
\method{print}{contOrd}(x, ...)
}
\arguments{
\item{formula}{A formula object (fixed effects).}
}
\description{
This function performs the comtinuous ordinal regression with logt link using the generalized logistic function as g function and without random effects.
}
\examples{
print()
}
\keyword{likelihood,}
\keyword{log-likelihood.}

