% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ocm.methods.R
\name{print.ocm}
\alias{print.ocm}
\title{Print continuous ordinal regression objects}
\usage{
\method{print}{ocm}(x, ...)
}
\arguments{
\item{x}{an object of class \code{ocm}, usually, a result of a call to \code{ocm}}

\item{...}{further arguments passed to or from other methods}
}
\value{
Prints an \code{ocm} object
}
\description{
\code{print.ocm} is the ordinalCont specific method for the generic function \code{print}, 
which prints objects of class \code{ocm}.
}
\author{
Maurizio Manuguerra, Gillian Heller
}
\seealso{
\code{\link{ocm}}, \code{\link{summary.ocm}}
}
\keyword{likelihood,}
\keyword{log-likelihood.}

