% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ocmm.methods.R
\name{print.ocmm}
\alias{print.ocmm}
\title{Print  a Continuous Ordinal Mixed Model Object}
\usage{
\method{print}{ocmm}(x, ...)
}
\arguments{
\item{x}{an object of class \code{ocmm}, usually, a result of a call to \code{ocmm}}

\item{...}{further arguments passed to or from other methods}
}
\value{
Prints an \code{ocmm} object
}
\description{
This function prints an \code{ocmm} object
}
\examples{
\dontrun{
fit.overall.rnd  <- ocmm(overall  ~ cycleno + age + bsa + treatment + (1|randno), data=ANZ0001)
print(fit.overall.rnd)
}
}
\author{
Maurizio Manuguerra, Gillian Heller
}
\seealso{
\code{\link{ocmm}}, \code{\link{summary.ocmm}}
}
\keyword{likelihood,}
\keyword{log-likelihood.}

