% Generated by roxygen2 (4.1.0.9000): do not edit by hand
% Please edit documentation in R/ocmm.methods.R
\name{print.predict.ocmm}
\alias{print.predict.ocmm}
\title{Print the output of predict method}
\usage{
\method{print}{predict.ocmm}(x, ...)
}
\arguments{
\item{x}{An object of class "predict.ocm".}

\item{...}{Further arguments passed to or from other methods.}
}
\description{
print method for class "predict.ocm"
}
\keyword{predict}

