% Generated by roxygen2 (4.1.0.9000): do not edit by hand
% Please edit documentation in R/functions.R
\name{print.summary.contOrd}
\alias{print.summary.contOrd}
\title{Summarizing Continuous Ordinal Fits}
\usage{
## S3 method for class 'contOrd'
summary(object, correlation = FALSE, symbolic.cor = FALSE, ...)
## S3 method for class 'summary.contOrd'
print(x, digits = max(3, getOption("digits") - 3),
     symbolic.cor = x$symbolic.cor,
     signif.stars = getOption("show.signif.stars"), ...)
}
\arguments{
\item{x}{An object of class "summary.contOrd", usually, a result of a call to summary.contOrd.}

\item{...}{Further arguments passed to or from other methods.}

\item{object}{An object of class "contOrd", usually, a result of a call to contOrd.}

\item{digits}{The number of significant digits to use when printing.}
}
\description{
summary method for class "contOrd"
}
\examples{
print.summary.contOrd()
}
\keyword{summary}

