% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ocmm.methods.R
\name{print.summary.ocmm}
\alias{print.summary.ocmm}
\title{Summarizing Continuous Ordinal Fits}
\usage{
\method{print}{summary.ocmm}(x, ...)
}
\arguments{
\item{x}{An object of class "summary.ocmm", usually, a result of a call to summary.ocmm.}

\item{...}{Further arguments passed to or from other methods.}
}
\description{
summary method for class ``summary.ocmm"
}
\keyword{summary}

