% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ordinalCont.R
\docType{data}
\name{qol}
\alias{qol}
\title{QoL scores marked by 80 patients at five time marks (only baseline included).}
\format{A data frame with 80 rows and 13 variables}
\usage{
data(qol)
}
\description{
A dataset containing the QoL scores of 80 patients at five time marks: baseline, 2, 4, 6, 12 months after a stroke event. The variables are as follows:
}
\details{
\itemize{
  \item ID. The ID number of the patient.
  \item Age Age of patient at entry in the study.
  \item Gender Sex of patient.
  \item Height
  \item Weight
  \item BMI
  \item Side.Affected
  \item Time...stroke duration of ...
  \item Sensation ...
  \item Neglect ...
  \item Spasticity ...
  \item Contracture ...
  \item M1.EQoL.health QoL score (from 0 to 100)
}
}
\keyword{datasets}

