% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ocm.methods.R
\name{summary.ocm}
\alias{summary.ocm}
\title{Summarizing Continuous Ordinal Fits}
\usage{
\method{summary}{ocm}(object, full = F, ...)
}
\arguments{
\item{object}{an object of class \code{ocm}, usually a result of a call to \code{ocm}}

\item{full}{logical, if TRUE (the default) all the parameters are printed; if FALSE, only the fixed effects are printed.}

\item{...}{further arguments passed to or from other methods}
}
\description{
Summary method for class \code{ocm}
}
\examples{
fit.overall  <- ocm(overall  ~ cycleno + age + bsa + treatment, data=ANZ0001.sub, scale=c(0,100))
summary(fit.overall)
}
\seealso{
\code{\link{ocm}}, \code{\link{print.ocm}}
}
\author{
Maurizio Manuguerra, Gillian Heller
}
\keyword{summary}
