% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ocmm.methods.R
\name{summary.ocmm}
\alias{summary.ocmm}
\title{Summarizing Continuous Ordinal Mixed Model Fits}
\usage{
\method{summary}{ocmm}(object, ...)
}
\arguments{
\item{object}{an object of class \code{ocmm}, usually a result of a call to \code{ocmm}}

\item{...}{further arguments passed to or from other methods}
}
\description{
Summary method for class \code{ocmm}
}
\examples{
\dontrun{
fit.overall.rnd  <- ocmm(overall  ~ cycleno + age + bsa + treatment + (1|randno), data=ANZ0001)
summary(fit.overall.rnd)
}
}
\author{
Maurizio Manuguerra, Gillian Heller
}
\keyword{summary}

