% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ocm.methods.R
\name{terms.ocm}
\alias{terms.ocm}
\title{Model Terms}
\usage{
\method{terms}{ocm}(x, random.terms = TRUE, ...)
}
\arguments{
\item{x}{an object of class \code{ocm}, usually, a result of a call to \code{ocm}.}

\item{random.terms}{a logical indicating if random terms have to be included in the terms object. Defaults to TRUE.}

\item{...}{further arguments passed to or from other methods.}
}
\value{
An object of class c("terms", "formula") which contains the terms representation of a symbolic model.
}
\description{
\code{terms.ocm} is the ordinalCont specific method for the generic function \code{terms}, 
which extracts model terms from objects of class \code{ocm}.
}
\author{
Maurizio Manuguerra, Gillian Heller
}
