% Generated by roxygen2 (4.1.0.9000): do not edit by hand
% Please edit documentation in R/functions.R
\name{test_function}
\alias{test_function}
\title{A Function}
\usage{
test_function(a = 0)
}
\arguments{
\item{a}{The param of the function, any number will do. Defaults to 0.}
}
\description{
This function allows you to …
}
\examples{
test_function()
}
\keyword{useless}

