% Generated by roxygen2 (4.1.0.9000): do not edit by hand
% Please edit documentation in R/ocm.methods.R
\name{vcov.ocm}
\alias{vcov.ocm}
\title{Calculate Variance-Covariance Matrix for a Fitted Model Object}
\usage{
\method{vcov}{ocm}(object, ...)
}
\arguments{
\item{object}{An ocm object.}

\item{...}{Further arguments to be passed to methods.}
}
\description{
Calculate Variance-Covariance Matrix for a Fitted Model Object
}

