% Generated by roxygen2 (4.1.0.9000): do not edit by hand
% Please edit documentation in R/functions.R
\name{anova.ocm}
\alias{anova.ocm}
\title{Anova method for Continuous Ordinal Fits}
\usage{
\method{anova}{ocm}(object, ...)
}
\arguments{
\item{...}{one or more additional ocm objects.}

\item{x}{An ocm object.}
}
\description{
Comparison of continuous ordinal models in likelihood ratio tests.
}
\keyword{anova}

