% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ocm.methods.R
\name{anova.ocm}
\alias{anova.ocm}
\title{Anova method for Continuous Ordinal Fits [GH up to here 19/3/15]}
\usage{
\method{anova}{ocm}(object, ...)
}
\arguments{
\item{object}{an \code{ocm} object}

\item{...}{one or more additional ocm objects.}
}
\description{
Comparison of continuous ordinal models in likelihood ratio tests.
THIS IS A TEST TO SEE IF UPDATES FROM MQ ARE WORKING
}
\keyword{anova}

