% Generated by roxygen2 (4.1.0.9000): do not edit by hand
% Please edit documentation in R/ocm.methods.R
\name{anova.ocm}
\alias{anova.ocm}
\title{Anova method for Continuous Ordinal Fits [GH up to here 19/3/15]}
\usage{
\method{anova}{ocm}(object, ...)
}
\arguments{
\item{object}{an object of class \code{ocm}}

\item{...}{one or more additional \code{ocm} objects.}
}
\description{
Comparison of continuous ordinal models in likelihood ratio tests.
}
\keyword{anova}

