% Generated by roxygen2 (4.1.0.9000): do not edit by hand
% Please edit documentation in R/ocmm.methods.R
\name{anova.ocmm}
\alias{anova.ocmm}
\title{Anova method for Continuous Ordinal Fits}
\usage{
\method{anova}{ocmm}(object, ...)
}
\arguments{
\item{object}{An ocm object.}

\item{...}{one or more additional ocm objects.}
}
\description{
Comparison of continuous ordinal models in likelihood ratio tests.
}
\examples{
\dontrun{
fitLaplace = ocmm(vas ~ lasert1+lasert2+lasert3+ (1|ID), data=pain, quad="Laplace")
anova(fitLaplace, update(fitLaplace, . ~ . + localisa))
}
}
\keyword{anova}

