% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ordinalCont.R
\docType{data}
\name{ANZ0001}
\alias{ANZ0001}
\title{ANZ0001 trial
The ANZ0001 trial, conducted by the ANZ Breast Cancer Trials Group, is an unblinded, multi-centre, randomised trial with three chemotherapy treatment arms concluded in 2005 (Stockler, Sour- jina, Grimison, Gebski, Byrne, Harvey, Francis, Nowak, Hazel, Forbes, and Group, 2007).
Health-related quality of life measures (Overall quality of life, Physical Well-Being, Mood, Pain, Nausea and Vomiting, Appetite) are assessed at each chemotherapy treatment cycle, from randomization until disease progression, when treatment is interrupted.
The treatments Intermittent Capecitabine (IC) and Continuous Capecitabine (CC) are compared with the standard combination treatment CMF, each with its own protocol.
There is no maximum duration of treatment, but it is interrupted on disease progression, or when patient intolerance or unacceptable toxicity are recorded.
The data set is extracted from the ANZ0001 trial and contains information from 292 patients with complete quality of life measurements.

The variables are as follows:

\itemize{
  \item randno. The ID number of the patient.
  \item cycleno The chemo cycle number.
  \item age Age of patient at entry in the study.
  \item bsa Patient Body Surface Area (m2).
  \item treatment The treatment received by the patient.
  \item overall Overall quality of life as recorded by the patient on a LASA scale, here normalised in the interval (0, 1).
  \item phys Physical Well-Being as recorded by the patient on a LASA scale, here normalised in the interval (0, 1).
  \item mood Mood as recorded by the patient on a LASA scale, here normalised in the interval (0, 1).
  \item pain Pain as recorded by the patient on a LASA scale, here normalised in the interval (0, 1).
  \item nausvom Nausea and Vomiting as recorded by the patient on a LASA scale, here normalised in the interval (0, 1).
  \item appetite Appetite as recorded by the patient on a LASA scale, here normalised in the interval (0, 1).
}}
\format{A data frame with 3473 rows and 11 variables}
\usage{
data(ANZ0001)
}
\description{
ANZ0001 trial
The ANZ0001 trial, conducted by the ANZ Breast Cancer Trials Group, is an unblinded, multi-centre, randomised trial with three chemotherapy treatment arms concluded in 2005 (Stockler, Sour- jina, Grimison, Gebski, Byrne, Harvey, Francis, Nowak, Hazel, Forbes, and Group, 2007).
Health-related quality of life measures (Overall quality of life, Physical Well-Being, Mood, Pain, Nausea and Vomiting, Appetite) are assessed at each chemotherapy treatment cycle, from randomization until disease progression, when treatment is interrupted.
The treatments Intermittent Capecitabine (IC) and Continuous Capecitabine (CC) are compared with the standard combination treatment CMF, each with its own protocol.
There is no maximum duration of treatment, but it is interrupted on disease progression, or when patient intolerance or unacceptable toxicity are recorded.
The data set is extracted from the ANZ0001 trial and contains information from 292 patients with complete quality of life measurements.

The variables are as follows:

\itemize{
  \item randno. The ID number of the patient.
  \item cycleno The chemo cycle number.
  \item age Age of patient at entry in the study.
  \item bsa Patient Body Surface Area (m2).
  \item treatment The treatment received by the patient.
  \item overall Overall quality of life as recorded by the patient on a LASA scale, here normalised in the interval (0, 1).
  \item phys Physical Well-Being as recorded by the patient on a LASA scale, here normalised in the interval (0, 1).
  \item mood Mood as recorded by the patient on a LASA scale, here normalised in the interval (0, 1).
  \item pain Pain as recorded by the patient on a LASA scale, here normalised in the interval (0, 1).
  \item nausvom Nausea and Vomiting as recorded by the patient on a LASA scale, here normalised in the interval (0, 1).
  \item appetite Appetite as recorded by the patient on a LASA scale, here normalised in the interval (0, 1).
}
}
\keyword{datasets}

