% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ocm.methods.R
\name{coef.ocm}
\alias{coef.ocm}
\title{Extract Model Coefficients}
\usage{
\method{coef}{ocm}(object, ...)
}
\arguments{
\item{object}{an object of class \code{ocm}, usually, a result of a call to \code{ocm}.}

\item{...}{further arguments passed to or from other methods.}
}
\value{
A named numeric vector with the coefficients extracted from the model object.
}
\description{
\code{coef.ocm} is the ordinalCont specific method for the generic function \code{coef}, 
which extracts model coefficients from objects of class \code{ocm}.
}
\author{
Maurizio Manuguerra, Gillian Heller
}
