% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ocm.methods.R
\name{coef.ocm}
\alias{coef.ocm}
\title{Extract Model Coefficients}
\usage{
coef(object, ...)
}
\arguments{
\item{object}{an \code{ocm} object}

\item{...}{other arguments}
}
\value{
vector of model coefficients in \code{object}
}
\description{
Extract Model Coefficients
}
\seealso{
\code{\link{ocm}}
}

