% Generated by roxygen2 (4.1.0.9000): do not edit by hand
% Please edit documentation in R/ocm.methods.R
\name{confint.ocm}
\alias{confint.ocm}
\title{title Confidence Intervals for Model Parameters}
\usage{
\method{coef}{ocm}(object, parm, level = 0.95)
}
\arguments{
\item{object}{a fitted model object}

\item{parm}{a specification of which parameters are to be given confidence intervals. If missing, all parameters are considered.}

\item{level}{the confidence level required.}

\item{...}{additional argument(s) for methods.}
}
\description{
title Confidence Intervals for Model Parameters
}
\examples{
fit <- ocm(vas ~ lasert1 + lasert2 + lasert3, data = pain)
confint(fit)
}

