% Generated by roxygen2 (4.1.0.9000): do not edit by hand
% Please edit documentation in R/utils.R
\name{dg_glf}
\alias{dg_glf}
\title{Derivative of generalized logistic g function}
\usage{
dg_glf(v, par)
}
\arguments{
\item{v}{Vector of standarized scores from the continuous ordinal scale.}

\item{par}{Vector of B, the slope of the curve, and T, the symmetry of the curve.}
}
\description{
This function compute a parametric version of the g function following Richards (1959): dg(v)/dv = T/B * 1/[v*(1-v^T)]
}
\keyword{Richards,}
\keyword{derivative,}
\keyword{function.}
\keyword{generalized}
\keyword{logistic}

