% Generated by roxygen2 (4.1.0.9000): do not edit by hand
% Please edit documentation in R/functions.R
\name{g_glf}
\alias{g_glf}
\title{Generalized logistic g function}
\usage{
g_glf(v, par)
}
\arguments{
\item{v}{Vector of standarized scores from the continuous ordinal scale.}

\item{par}{Vector of M, the offset of the curve; B, the slope of the curve, and T, the symmetry of the curve.}
}
\description{
This function compute a parametric version of the g function following Richards (1959): g(v) = M + 1/B * log(Tv^T/(1-v^T))
}
\examples{
g_glf()
}
\keyword{Richards,}
\keyword{function.}
\keyword{generalized}
\keyword{logistic}

