% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/utils.R
\name{g_glf}
\alias{g_glf}
\title{Generalized logistic g function}
\usage{
g_glf(v, par)
}
\arguments{
\item{v}{vector of standardized scores from the continuous ordinal scale, 0<v<1.}

\item{par}{vector of 2 elements: B, the slope of the curve, and T, the symmetry of the curve.}
}
\value{
A vector of length equal to the length of v, with values g(v).
}
\description{
This function computes a parametric version of the g function following Richards (1959):
\deqn{g(v) = M + \frac{1}{B} \log\left(\frac{Tv^T}{1-v^T}\right)}
 M is omitted as an intercept is always fitted.
}
\details{
The generalized logistic functions maps from (0,1) to \eqn{(-\infty,\infty)}.
B is the slope of the curve,  T is the symmetry and M is the offset.
M is absorbed into the intercept of the model and so is set to zero in this function.
}
\references{
Richards, F. (1959). A flexible growth function for empirical use,
\emph{Journal of Experimental Botany}, 10, 290–301.
}
\keyword{Richards,}
\keyword{function.}
\keyword{generalized}
\keyword{logistic}

