% Generated by roxygen2 (4.1.0.9000): do not edit by hand
% Please edit documentation in R/utils.R
\name{g_glf_inv}
\alias{g_glf_inv}
\title{Inverse of generalized logistic g function}
\usage{
g_glf_inv(W, par)
}
\arguments{
\item{W}{Vector of scores on the latent scale - M, the offset of the g function, estimated as the intercept of the model. W=-x'B [without intercept] = -x'B - M [as the model is estimated with an intercept].}

\item{par}{Vector of 2 elements: B, the slope of the curve, and T, the symmetry of the curve.}
}
\description{
This function computes the inverse of a parametric version of the g function following Richards (1959): \deqn{g(v) = M + \frac{1}{B} log\left(\frac{Tv^T}{1-v^T}\right)}. M is omitted as an intercept is always fitted.
}
\keyword{Richards,}
\keyword{function.}
\keyword{generalized}
\keyword{logistic}

