% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ocm.methods.R
\name{logLik.ocm}
\alias{logLik.ocm}
\title{Extract the Log-Likelihood}
\usage{
\method{logLik}{ocm}(object, ...)
}
\arguments{
\item{object}{an \code{ocm} object.}

\item{...}{Further arguments to be passed to methods.}
}
\value{
Returns the log-likelihood of an \code{ocm} object
}
\description{
Extract the Log-Likelihood
}
\examples{
ANZ0001.ocm <- ANZ0001[ANZ0001$cycleno==0 | ANZ0001$cycleno==5,]
ANZ0001.ocm$cycleno[ANZ0001.ocm$cycleno==5] <- 1
fit.overall  <- ocm(overall  ~ cycleno + age + bsa + treatment, data=ANZ0001.ocm)
logLik(fit.overall)
}
\seealso{
\code{\link{ocm}}
}

