% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ocm.methods.R
\name{logLik.ocm}
\alias{logLik.ocm}
\title{Extract Log-Likelihood}
\usage{
logLik(object, ...)
}
\arguments{
\item{object}{an \code{ocm} object.}
}
\value{
log likelihood of \code{ocm} object
}
\description{
Extract Log-Likelihood
}
\seealso{
\code{\link{ocm}}
}

