% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ocm.methods.R
\name{model.frame.ocm}
\alias{model.frame.ocm}
\title{Extract the Model Frame from a Fit}
\usage{
\method{model.frame}{ocm}(object, ...)
}
\arguments{
\item{object}{An ocm object.}

\item{...}{Further arguments to be passed to methods.}
}
\value{
model frame
}
\description{
Extract the Model Frame from a Fit
}
\seealso{
\code{\link{ocm}}
}

