% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ocm.methods.R
\name{model.frame.ocm}
\alias{model.frame.ocm}
\title{Extracting the Model Frame from a Fit}
\usage{
\method{model.frame}{ocm}(object, ...)
}
\arguments{
\item{object}{An ocm object.}

\item{...}{Further arguments to be passed to methods.}
}
\description{
Extracting the Model Frame from a Fit
}

