% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ocm.methods.R
\name{model.matrix.ocm}
\alias{model.matrix.ocm}
\title{Extract the Model Matrix from a Fit}
\usage{
\method{model.matrix}{ocm}(object, ...)
}
\arguments{
\item{object}{an \code{ocm} object}

\item{...}{Further arguments to be passed to methods.}
}
\value{
model matrix
}
\description{
Extract the Model Matrix from a Fit
}
\seealso{
\code{\link{ocm}}
}

