% Generated by roxygen2 (4.1.0.9000): do not edit by hand
% Please edit documentation in R/ocm.R
\name{negloglik_glf}
\alias{negloglik_glf}
\title{Log-likelihood function for the fixed-effects model, using the generalized logistic
function as g function and the logit link function}
\usage{
negloglik_glf(par, v, d.matrix, wts, len_beta)
}
\arguments{
\item{par}{vector of regression coefficients (first \code{len_beta} elements),
and \code{M},  \code{B}, \code{T}, (offset, slope and symmetry of the g function -
last 3 elements)}

\item{v}{vector of standardized scores from the continuous ordinal scale}

\item{d.matrix}{design matrix (fixed effects)}

\item{wts}{????}

\item{len_beta}{length of the regression coefficients vector}
}
\value{
minus the log-likelihood at parameter values \code{par}
}
\description{
Log-likelihood function for the fixed-effects model, using the generalized logistic
function as g function and the logit link function
}
\details{
This function computes minus the log-likelihood function for a fixed-effects model using the
generalized logistic function as g function and the logit link function.
}
\keyword{likelihood,}
\keyword{log-likelihood.}

