% Generated by roxygen2 (4.1.0.9000): do not edit by hand
% Please edit documentation in R/functions.R
\name{negloglik_glf}
\alias{negloglik_glf}
\title{Log-likelihood function for single subject using the generalized logistic function as g function and without random effects}
\usage{
negloglik_glf(par, v, x, len_beta)
}
\arguments{
\item{v}{Vector of standarized scores from the continuous ordinal scale.}

\item{x}{Design matrix (fixed effects).}

\item{beta}{Regression coefficients.}

\item{M}{The offset of the curve.}

\item{B}{The slope of the curve.}

\item{T}{The symmetry of the curve.}
}
\description{
This function compute the Log-likelihood function for single subject using the generalized logistic function as g function and without random effects.
}
\examples{
negloglik_glf()
}
\keyword{likelihood,}
\keyword{log-likelihood.}

