% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ocm.R
\name{negloglik_glf}
\alias{negloglik_glf}
\title{Log-likelihood function for the fixed-effects model, using the generalized logistic function as
g function and the logit link function.}
\usage{
negloglik_glf(par, v, d.matrix, len_beta)
}
\arguments{
\item{par}{Vector of B, the slope of the curve, and T, the symmetry of the curve.}

\item{v}{Vector of standarized scores from the continuous ordinal scale.}

\item{d.matrix}{Design matrix (fixed effects).}

\item{len_beta}{Length of the regression coefficients vector.}
}
\description{
This function compute the log-likelihood function for a fixed-effects model using the
generalized logistic function as g function and the logit link function.
}
\keyword{likelihood,}
\keyword{log-likelihood.}

