% Generated by roxygen2 (4.1.0.9000): do not edit by hand
% Please edit documentation in R/ocm.methods.R
\name{nobs.ocm}
\alias{nobs.ocm}
\title{Extract the Number of Observations from a Fit}
\usage{
\method{nobs}{ocm}(object, ...)
}
\arguments{
\item{object}{an \code{ocm} object}

\item{...}{Further arguments to be passed to methods.}
}
\value{
number of observations
}
\description{
Extract the Number of Observations from a Fit
}
\seealso{
\code{\link{ocm}}
}

