% Generated by roxygen2 (4.1.0.9000): do not edit by hand
% Please edit documentation in R/functions.R
\name{ocm}
\alias{ocm}
\title{Continuous ordinal regression}
\usage{
ocm(x, ...)
}
\arguments{
\item{formula}{A formula object (fixed effects).}
}
\description{
This function performs the continuous ordinal regression with logit link using the generalized logistic function as g function and without random effects.
}
\examples{
ocm()
}
\keyword{likelihood,}
\keyword{log-likelihood.}

