% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/functions.R
\name{ocm}
\alias{ocm}
\title{Ordinal regression for continuous scales}
\usage{
ocm(formula, data, start = NULL, control = list(), link = c("logit"),
  gfun = c("glf"), ...)
}
\arguments{
\item{formula}{a formula expression as for regression models, of the form response ~ predictors. Only fixed effects are supported. The model must have an intercept: attempts to remove one will lead to a warning and will be ignored (TODO).}

\item{data}{an optional data frame in which to interpret the variables occurring in the formulas.}

\item{start}{initial values for the parameters in the format c(alpha, beta, zeta), where alpha are the threshold parameters (adjusted for potential nominal effects), beta are the regression parameters and zeta are the scale parameters. (CHANGETHIS)}

\item{control}{a list of control parameters passed on to clm.control.}

\item{link}{link function, i.e., the type of location-scale distribution assumed for the latent distribution. The default "logit" link gives the proportional odds model.}

\item{gfun}{A smooth monotonic function capable of capturing the non-linear nature of the ordinal measure. It defaults to the generalized logistic function, which is currently the only possibility.}

\item{...}{additional arguments are passed on to clm.control.}
}
\description{
This function performs the continuous ordinal regression with logit link using the generalized logistic function as g function and without random effects.
}
\examples{
# Change data set
fit = ocm(vas ~ lasert1+lasert2+lasert3, data=pain)
}
\keyword{likelihood,}
\keyword{log-likelihood,}
\keyword{ordinal}
\keyword{regression.}

