% Generated by roxygen2 (4.1.0.9000): do not edit by hand
% Please edit documentation in R/ocm.R
\name{ocm}
\alias{ocm}
\title{Ordinal regression for continuous scales}
\usage{
ocm(formula, data, weights, start = NULL, control = list(),
  link = c("logit"), gfun = c("glf"), ...)
}
\arguments{
\item{formula}{a formula expression as for regression models, of the form
response ~ predictors. Only fixed effects are supported.
The model must have an intercept: attempts to remove one will lead to a warning and will be
ignored (TODO).}

\item{data}{an optional data frame in which to interpret the variables occurring in the formulas.}

\item{start}{initial values for the parameters in the format c(alpha, beta, zeta), where
alpha are the threshold parameters (adjusted for potential nominal effects), beta are the
regression parameters and zeta are the scale parameters. (CHANGETHIS)}

\item{control}{a list of control parameters passed on to clm.control.}

\item{link}{link function, i.e. the type of location-scale distribution assumed for the latent
distribution. The default "logit" link gives the proportional odds model.}

\item{gfun}{A smooth monotonic function capable of capturing the non-linear nature of the ordinal
measure. It defaults to the generalized logistic function, which is currently the only possibility.}

\item{...}{additional arguments are passed on to clm.control.}
}
\value{
an object of type \code{ocm}. Parameter estimates are in \code{coefficients}.
The last 3 elements of \code{coefficients} are the parameters of the g function:
\code{M},  \code{B},  and \code{T}.
}
\description{
This function performs continuous ordinal regression with logit link using the
generalized logistic function as g function. Random effects are not supported.
}
\details{
Ordinal regression analysis is a convenient tool for analyzing ordinal response variables
in the presence of covariates. We extend this methodology to the case of continuous self-rating
scales such as the Visual Analog Scale (VAS) used in pain assessment, or the Linear Analog
Self-Assessment (LASA) scales in quality of life studies. Subjects are
typically given a linear scale of 100 mm and asked to put a mark where they perceive
themselves. These scales  measure subjects'
perception of an intangible quantity, and cannot be handled as ratio variables because of their
inherent nonlinearity.  We express  the likelihood in terms of a function (the "g function")
 connecting the
scale with an underlying continuous latent  variable. In the current version the g function is taken as
the generalized logistic function (Richards 1959). This has 3 parameters:
 \code{M}, the offset, \code{B}, the slope, and \code{T}, the symmetry of the curve.
The link function is the inverse of the CDF of the assumed underlying distribution of the latent variable. Currently
the logit link, which corresponds to a standard logistic distribution, is implemented.
(This implies a proportional odds model.)
A regression framework supporting fixed effects
 is implemented. The likelihood is maximized using XXXXX MAURIZIO which function?.
}
\examples{
# Change data set
#fit <- ocm(vas ~ lasert1+lasert2+lasert3, data=pain)
ANZ0001.ocm <- ANZ0001[ANZ0001$cycleno==0 | ANZ0001$cycleno==5,]
ANZ0001.ocm$cycleno[ANZ0001.ocm$cycleno==5] <- 1
fit.overall  <- ocm(overall  ~ cycleno + age + bsa + treatment, data=ANZ0001.ocm)
fit.phys 	  <- ocm(phys 	  ~ cycleno + age + bsa + treatment, data=ANZ0001.ocm)
fit.pain 	  <- ocm(pain 	  ~ cycleno + age + bsa + treatment, data=ANZ0001.ocm)
fit.mood 	  <- ocm(mood 	  ~ cycleno + age + bsa + treatment, data=ANZ0001.ocm)
fit.nausvom  <- ocm(nausvom  ~ cycleno + age + bsa + treatment, data=ANZ0001.ocm)
fit.appetite <- ocm(appetite ~ cycleno + age + bsa + treatment, data=ANZ0001.ocm)
summary(fit.overall)
summary(fit.phys)
summary(fit.pain)
summary(fit.mood)
summary(fit.nausvom)
summary(fit.appetite)
par(mfrow=c(2,3))
plot(fit.overall, CIs='vcov', R=100)
plot(fit.phys, CIs='vcov', R=100)
plot(fit.pain, CIs='vcov', R=100)
plot(fit.mood, CIs='vcov', R=100)
plot(fit.nausvom, CIs='vcov', R=100)
plot(fit.appetite, CIs='vcov', R=100)
par(mfrow=c(1,1))
}
\author{
Maurizio Manuguerra, Gillian Heller
}
\references{
Manuguerra M, Heller GZ (2010). Ordinal Regression Models for Continuous Scales,
 \emph{The International Journal of Biostatistics}: 6(1), Article 14.

Richards, F. (1959). A flexible growth function for empirical use,
\emph{Journal of Experimental Botany}, 10, 290-301.
}
\seealso{
\code{\link{ocmm}}
}
\keyword{likelihood,}
\keyword{log-likelihood,}
\keyword{ordinal}
\keyword{regression.}

