% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/functions.R
\name{plot.ocm}
\alias{plot.ocm}
\title{Plot method for Continuous Ordinal Fits}
\usage{
\method{plot}{ocm}(x, ...)
}
\arguments{
\item{x}{An ocm object.}

\item{...}{Further arguments passed to or from other methods.}
}
\description{
Plot based on ocm object.
}
\keyword{plot}

