% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/functions.R
\name{plot.ocm}
\alias{plot.ocm}
\title{Plot method for Continuous Ordinal Fits}
\usage{
\method{plot}{ocm}(x, CIs = FALSE, ...)
}
\arguments{
\item{x}{An ocm object.}

\item{CIs}{A logical factor indicating if confidence inteval for the g function should be computed. This is done with bootstapping method and the operation could be lengthy.}

\item{...}{Further arguments passed to or from other methods.}
}
\description{
This function plots the g function as fitted in an ocm call.
}
\keyword{plot}

