% Generated by roxygen2 (4.1.0.9000): do not edit by hand
% Please edit documentation in R/functions.R
\name{plot.ocm}
\alias{plot.ocm}
\title{Plot method for Continuous Ordinal Fits}
\usage{
\method{plot}{ocm}(object, ...)
}
\arguments{
\item{object}{An ocm object.}
}
\description{
Plot based on ocm object.
}
\keyword{plot}

