% Generated by roxygen2 (4.1.0.9000): do not edit by hand
% Please edit documentation in R/functions.R
\name{plot.ocm}
\alias{plot.ocm}
\title{Plot method for Continuous Ordinal Fits}
\usage{
\method{plot}{ocm}(x, CIs = c("simple", "rnd-x-bootstrap", "fix-x-bootstrap"),
  R = 1000, ...)
}
\arguments{
\item{x}{An ocm object.}

\item{CIs}{Indicates if confidence bands for the g function should be computed based on the Wald 95\% CIs or by bootstrapping.}

\item{...}{Further arguments passed to or from other methods.}
}
\description{
This function plots the g function as fitted in an ocm call.
}
\keyword{plot}

