% Generated by roxygen2 (4.1.0.9000): do not edit by hand
% Please edit documentation in R/ocm.methods.R
\name{plot.ocm}
\alias{plot.ocm}
\title{Plot method for Continuous Ordinal Fits}
\usage{
\method{plot}{ocm}(x, CIs = c("no", "vcov", "rnd.x.bootstrap",
  "fix.x.bootstrap", "param.bootstrap"), R = 1000, ...)
}
\arguments{
\item{x}{an object of class \code{ocm}}

\item{CIs}{method used for confidence bands for the g function. \code{"no"} = no CIS; \code{"vcov"} = Wald;
\code{"rnd.x.bootstrap"} = random-x bootstrap; \code{"fix.x.bootstrap"} = bootstrap with fixed-x
resampling; \code{"param.bootstrap"} = parametric bootstrap.}

\item{R}{the number of bootstrap replicates [ignored if CIs='no']}

\item{...}{further arguments passed to or from other methods}
}
\description{
Plots the g function as fitted in an \code{ocm} call.
}
\details{
The fitted g function of an \code{ocm} object is plotted.
If \code{CIs} is not \code{"no"}, 95\% confidence bands are also plotted.
Confidence bands computed with any of the bootstrapping options are
obtained with simple percentiles.
}
\examples{
fit <- ocm(vas ~ lasert1 + lasert2 + lasert3, data = pain)
plot(fit, CIs="vcov")
}
\author{
Maurizio Manuguerra
}
\seealso{
\code{\link{ocm}}
}
\keyword{plot}

