% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ocm.methods.R
\name{plot.ocm}
\alias{plot.ocm}
\title{Plot method for Continuous Ordinal Fits}
\usage{
\method{plot}{ocm}(x, CIs = c("var", "vcov", "rnd.x.bootstrap",
  "fix.x.bootstrap", "param.bootstrap"), R = 1000, ...)
}
\arguments{
\item{x}{An ocm object.}

\item{CIs}{Indicates if confidence bands for the g function should be computed based on the Wald 95\% CIs or by bootstrapping. In  the latter case, bootstrapping can be performed using a random-x or a fixed-x resampling. 95\% CIs computed with either of the bootstrapping options are obtained with simple percentiles.}

\item{R}{The number of bootstrap replicates.}

\item{...}{Further arguments passed to or from other methods.}
}
\description{
This function plots the g function as fitted in an ocm call.
}
\keyword{plot}

