% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ocm.methods.R
\name{plot.predict.ocm}
\alias{plot.predict.ocm}
\title{Plot  probability densities  from  output of  predict method}
\usage{
\method{plot}{predict.ocm}(x, records = NULL, ...)
}
\arguments{
\item{x}{An object of class \code{predict.ocm}}

\item{records}{An integer or a vector of integers. The number of the record/s
in the data set for which the density has to be plotted. If not specified, the
function will  plot all records.}

\item{...}{further arguments passed to or from other methods}
}
\description{
Plot method for class \code{predict.ocm}
}
\details{
The probability densities from \code{predict.ocm}  are plotted.
}
\examples{
ANZ0001.ocm <- ANZ0001[ANZ0001$cycleno==0 | ANZ0001$cycleno==5,]
ANZ0001.ocm$cycleno[ANZ0001.ocm$cycleno==5] <- 1
fit.overall <- ocm(overall ~ cycleno + age + bsa + treatment, data=ANZ0001.ocm)
pred <- predict(fit.overall)
plot(pred)
}
\author{
Maurizio Manuguerra, Gillian Heller
}
\seealso{
\code{\link{predict.ocm}}, \code{\link{ocm}}
}
\keyword{plot}
\keyword{predict,}

