% Generated by roxygen2 (4.1.0.9000): do not edit by hand
% Please edit documentation in R/functions.R
\name{predict.ocm}
\alias{predict.ocm}
\title{Predict method for Continuous Ordinal Fits}
\usage{
\method{predict}{ocm}(object, newdata = NULL, ...)
}
\arguments{
\item{love}{Do you love cats? Defaults to TRUE.}
}
\description{
Predicted values based on ocm object.
}
\examples{
cat_function()
}
\keyword{cats}

