% Generated by roxygen2 (4.1.0.9000): do not edit by hand
% Please edit documentation in R/functions.R
\name{predict.ocm}
\alias{predict.ocm}
\title{Predict method for Continuous Ordinal Fits}
\usage{
\method{predict}{ocm}(object, newdata = NULL, ...)
}
\arguments{
\item{object}{An ocm object.}

\item{newdata}{optionally, a data frame in which to look for variables with which to predict. Note that all predictor variables should be present having the same names as the variables used to fit the model.}
}
\description{
Predicted values based on ocm object.
}
\keyword{predict}

