% Generated by roxygen2 (4.1.0.9000): do not edit by hand
% Please edit documentation in R/ocm.methods.R
\name{predict.ocm}
\alias{predict.ocm}
\title{Predict method for Continuous Ordinal Fits}
\usage{
\method{predict}{ocm}(object, newdata = NULL, ...)
}
\arguments{
\item{object}{an object of class \code{"ocm"}, usually, a result of a call to \code{ocm}}

\item{newdata}{optionally, a data frame in which to look for variables with
which to predict.
Note that all predictor variables should be present, having the same names as the variables
used to fit the model. If \code{NULL}, predictions are computed for the original dataset.}

\item{...}{Further arguments passed to or from other methods.}
}
\value{
MAURIZIO we need to specify this (I'm not sure what you've done)
}
\description{
Predicted values based on \code{ocm} object.
}
\details{
MAURIZIO we need to specify this (I'm not sure what you've done)
}
\examples{
xxx add example
}
\seealso{
\code{\link{ocm}}
}
\keyword{predict}

