% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ocm.methods.R
\name{predict.ocm}
\alias{predict.ocm}
\title{Predict method for Continuous Ordinal Fits}
\usage{
\method{predict}{ocm}(object, newdata = NULL, ...)
}
\arguments{
\item{object}{an object of class \code{"ocm"}, usually a result of a call to \code{ocm}}

\item{newdata}{optionally, a data frame in which to look for variables with
which to predict.
Note that all predictor variables should be present, having the same names as the variables
used to fit the model. If \code{NULL}, predictions are computed for the original dataset.}

\item{...}{Further arguments passed to or from other methods.}
}
\value{
A list containing the following components:
\tabular{ll}{
\code{mode} \tab a vector of length equal to the number of observations. Each element is the mode of v,
the ordinal continuous random variable, conditional on the covariates in the model\cr\cr
\code{density} \tab a matrix with number of rows equal to the number of observations. Each row
contains the values of the density function of v conditional on the covariates in the model.
The density function is calculated over 100 equally-spaced values of v in (0,1)\cr\cr
\code{x} \tab a vector with the 100 equally-spaced values of v in (0,1) used to compute the density of v\cr\cr
\code{formula} \tab the formula used to fit the model\cr\cr
\code{newdata}\tab a new data frame used to make predictions. It takes value NULL if no new data frame has been used.
}
}
\description{
Predicted values based on \code{ocm} object.
}
\details{
MAURIZIO we need to specify this (I'm not sure what you've done)
}
\examples{
ANZ0001.ocm <- ANZ0001[ANZ0001$cycleno==0 | ANZ0001$cycleno==5,]
ANZ0001.ocm$cycleno[ANZ0001.ocm$cycleno==5] <- 1
fit.overall <- ocm(overall ~ cycleno + age + bsa + treatment, data=ANZ0001.ocm)
pred <- predict(fit.overall)
plot(pred)
}
\author{
Maurizio Manuguerra
}
\seealso{
\code{\link{ocm}}
}
\keyword{predict}

