% Generated by roxygen2 (4.1.0.9000): do not edit by hand
% Please edit documentation in R/ocm.methods.R
\name{print.anova.ocm}
\alias{print.anova.ocm}
\title{Print anova.ocm objects}
\usage{
\method{print}{anova.ocm}(x, digits = max(getOption("digits") - 2, 3),
  signif.stars = getOption("show.signif.stars"), ...)
}
\arguments{
\item{x}{An object of class "anova.ocm".}

\item{...}{Further arguments passed to or from other methods.}
}
\description{
Print the results of the comparison of continuous ordinal models in likelihood ratio tests.
}
\keyword{anova}
\keyword{summary,}

