% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ocmm.methods.R
\name{print.anova.ocmm}
\alias{print.anova.ocmm}
\title{Print anova.ocm objects}
\usage{
\method{print}{anova.ocmm}(x, digits = max(getOption("digits") - 2, 3),
  signif.stars = getOption("show.signif.stars"), ...)
}
\arguments{
\item{x}{An object of class ``anova.ocmm".}

\item{digits}{controls the number of digits to print. Defaults to the maximum between the value returned by (getOption("digits") - 2) and 3}

\item{signif.stars}{a logical. Should the significance stars be printed? Defaults to the value returned by getOption("show.signif.stars")}

\item{...}{Further arguments passed to or from other methods.}
}
\value{
Prints \code{anova.ocmm} object
}
\description{
Print the results of the comparison of continuous ordinal models in likelihood ratio tests.
}
\keyword{anova}
\keyword{summary,}

