% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ocm.methods.R
\name{print.ocm}
\alias{print.ocm}
\title{Print continuous ordinal regression objects}
\usage{
\method{print}{ocm}(x, ...)
}
\arguments{
\item{x}{an object of class \code{ocm}, usually, a result of a call to \code{ocm}}

\item{...}{further arguments passed to or from other methods}
}
\value{
Prints an \code{ocm} object
}
\description{
\code{print.ocm} is the ordinalCont specific method for the generic function \code{print},
which prints objects of class \code{ocm}.
}
\examples{
ANZ0001.ocm <- ANZ0001[ANZ0001$cycleno==0 | ANZ0001$cycleno==5,]
ANZ0001.ocm$cycleno[ANZ0001.ocm$cycleno==5] <- 1
fit.overall  <- ocm(overall  ~ cycleno + age + bsa + treatment, data=ANZ0001.ocm)
print(fit.overall)
}
\author{
Maurizio Manuguerra, Gillian Heller
}
\seealso{
\code{\link{ocm}}, \code{\link{summary.ocm}}
}
\keyword{likelihood,}
\keyword{log-likelihood.}

