% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/functions.R
\name{print.ocm}
\alias{print.ocm}
\title{Continuous ordinal regression}
\usage{
\method{print}{ocm}(x, ...)
}
\arguments{
\item{x}{An object of class "ocm", usually, a result of a call to ocm.}

\item{...}{Further arguments passed to or from other methods.}
}
\description{
This function performs the comtinuous ordinal regression with logt link using the generalized logistic function as g function and without random effects.
}
\keyword{likelihood,}
\keyword{log-likelihood.}

