% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ocm.methods.R
\name{print.ocm}
\alias{print.ocm}
\title{Print continuous ordinal regression objects}
\usage{
\method{print}{ocm}(x, ...)
}
\arguments{
\item{x}{An object of class \code{'ocm'}, usually, a result of a call to \code{ocm}.}

\item{...}{Further arguments passed to or from other methods.}
}
\value{
Prints an \code{ocm} object
}
\description{
\code{print.ocm} is the ordinalCont specific method for the generic function \code{print},
which prints objects of class \code{'ocm'}.
}
\seealso{
\code{\link{ocm}}, \code{\link{summary.ocm}}
}
\keyword{likelihood,}
\keyword{log-likelihood.}

