% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ocmm.methods.R
\name{print.ocmm}
\alias{print.ocmm}
\title{Print continuous ordinal regression objects}
\usage{
\method{print}{ocmm}(x, ...)
}
\arguments{
\item{x}{An object of class "ocm", usually, a result of a call to ocm.}

\item{...}{Further arguments passed to or from other methods.}
}
\description{
This function prints an ocmm object
}
\keyword{likelihood,}
\keyword{log-likelihood.}

