% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ocm.methods.R
\name{print.predict.ocm}
\alias{print.predict.ocm}
\title{Print the output of predict method}
\usage{
\method{print}{predict.ocm}(x, ...)
}
\arguments{
\item{x}{an object of class \code{predict.ocm}}

\item{...}{further arguments passed to or from other methods}
}
\description{
Print method for class \code{predict.ocm}
}
\details{
The table of predictions from \code{predict.ocm} is printed.
}
\examples{
ANZ0001.ocm <- ANZ0001[ANZ0001$cycleno==0 | ANZ0001$cycleno==5,]
ANZ0001.ocm$cycleno[ANZ0001.ocm$cycleno==5] <- 1
fit.overall <- ocm(overall ~ cycleno + age + bsa + treatment, data=ANZ0001.ocm)
pred <- predict(fit.overall)
print(pred)
}
\author{
Maurizio Manuguerra, Gillian Heller
}
\seealso{
\code{\link{predict.ocm}}, \code{\link{ocm}}
}
\keyword{predict}

