% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ocm.methods.R
\name{print.predict.ocm}
\alias{print.predict.ocm}
\title{Print the output of predict method}
\usage{
\method{print}{predict.ocm}(x, ...)
}
\arguments{
\item{x}{An object of class \code{"predict.ocm"}}

\item{...}{Further arguments passed to or from other methods}
}
\description{
Print method for class \code{"predict.ocm"}
}
\details{
The table of predictions from \code{predict.ocm} is printed.
}
\seealso{
\code{\link{predict.ocm}}, \code{\link{ocm}}
}
\keyword{predict}

