% Generated by roxygen2 (4.1.0.9000): do not edit by hand
% Please edit documentation in R/functions.R
\name{print.summary.ocm}
\alias{print.summary.ocm}
\title{Summarizing Continuous Ordinal Fits}
\usage{
\method{print}{summary.ocm}(x, digits = getOption("digits"), ...)
}
\arguments{
\item{x}{An object of class "summary.ocm", usually, a result of a call to summary.ocm.}

\item{digits}{The number of significant digits to use when printing.}

\item{...}{Further arguments passed to or from other methods.}
}
\description{
summary method for class "ocm"
}
\keyword{summary}

