% Generated by roxygen2 (4.1.0.9000): do not edit by hand
% Please edit documentation in R/ocm.methods.R
\name{print.summary.ocm}
\alias{print.summary.ocm}
\title{Summarizing Continuous Ordinal Fits}
\usage{
\method{print}{summary.ocm}(x, ...)
}
\arguments{
\item{x}{An object of class \code{"summary.ocm"}, usually a result of a call to \code{summary.ocm}.}

\item{...}{further arguments passed to or from other methods.}
}
\description{
Summary method for class \code{"summary.ocm"}
}
\details{
The table of parameter estimates is printed.
}
\examples{
#Example here
}
\author{
Maurizio Manuguerra
}
\keyword{summary}

