% Generated by roxygen2 (4.1.0.9000): do not edit by hand
% Please edit documentation in R/functions.R
\name{summary.ocm}
\alias{summary.ocm}
\title{Summarizing Continuous Ordinal Fits}
\usage{
\method{summary}{ocm}(object, ...)
}
\arguments{
\item{object}{An object of class "ocm", usually, a result of a call to ocm.}

\item{...}{Further arguments passed to or from other methods.}
}
\description{
summary method for class "ocm"
}
\keyword{summary}

