% Generated by roxygen2 (4.1.0.9000): do not edit by hand
% Please edit documentation in R/functions.R
\name{summary.ocm}
\alias{summary.ocm}
\title{Summarizing Continuous Ordinal Fits}
\usage{
## S3 method for class 'ocm'
summary(object, correlation = FALSE, symbolic.cor = FALSE, ...)
## S3 method for class 'summary.ocm'
print(x, digits = max(3, getOption("digits") - 3),
     symbolic.cor = x$symbolic.cor,
     signif.stars = getOption("show.signif.stars"), ...)
}
\arguments{
\item{object}{An object of class "ocm", usually, a result of a call to ocm.}

\item{...}{Further arguments passed to or from other methods.}

\item{x}{An object of class "summary.ocm", usually, a result of a call to summary.ocm.}

\item{digits}{The number of significant digits to use when printing.}
}
\description{
summary method for class "ocm"
}
\keyword{summary}

