% Generated by roxygen2 (4.1.0.9000): do not edit by hand
% Please edit documentation in R/ocm.methods.R
\name{summary.ocm}
\alias{summary.ocm}
\title{Summarizing Continuous Ordinal Fits}
\usage{
\method{summary}{ocm}(object, ...)
}
\arguments{
\item{object}{An object of class \code{"ocm"}, usually, a result of a call to \code{ocm}.}

\item{...}{Further arguments passed to or from other methods.}
}
\description{
Summary method for class \code{'ocm'}
}
\examples{
xxx add example
}
\seealso{
\code{\link{ocm}}, \code{\link{print.ocm}}
}
\keyword{summary}

