% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ocm.methods.R
\name{summary.ocm}
\alias{summary.ocm}
\title{Summarizing Continuous Ordinal Fits}
\usage{
\method{summary}{ocm}(object, ...)
}
\arguments{
\item{object}{An object of class \code{ocm}, usually, a result of a call to \code{ocm}.}

\item{...}{Further arguments passed to or from other methods.}
}
\description{
Summary method for class \code{ocm}
}
\examples{
ANZ0001.ocm <- ANZ0001[ANZ0001$cycleno==0 | ANZ0001$cycleno==5,]
ANZ0001.ocm$cycleno[ANZ0001.ocm$cycleno==5] <- 1
fit.overall  <- ocm(overall  ~ cycleno + age + bsa + treatment, data=ANZ0001.ocm)
summary(fit.overall)
}
\author{
Maurizio Manuguerra
}
\seealso{
\code{\link{ocm}}, \code{\link{print.ocm}}
}
\keyword{summary}

