% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ocmm.methods.R
\name{summary.ocmm}
\alias{summary.ocmm}
\title{Summarizing Continuous Ordinal Fits}
\usage{
\method{summary}{ocmm}(object, ...)
}
\arguments{
\item{object}{An object of class "ocm", usually, a result of a call to ocm.}

\item{...}{Further arguments passed to or from other methods.}
}
\description{
summary method for class "ocm"
}
\keyword{summary}

