% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ocm.methods.R
\name{terms.ocm}
\alias{terms.ocm}
\title{Model Terms}
\usage{
\method{terms}{ocm}(object, ...)
}
\arguments{
\item{object}{An \code{ocm} object.}

\item{...}{Further arguments to be passed to methods.}
}
\value{
model terms
}
\description{
Model Terms
}
\seealso{
\code{\link{ocm}}
}

