% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ocm.methods.R
\name{vcov.ocm}
\alias{vcov.ocm}
\title{Calculate Variance-Covariance Matrix for a Fitted Model Object}
\usage{
\method{vcov}{ocm}(object, ...)
}
\arguments{
\item{object}{An \code{ocm} object.}

\item{...}{Further arguments to be passed to methods}
}
\value{
Variance-covariance matrix of model parameters
}
\description{
Calculate Variance-Covariance Matrix for a Fitted Model Object
}
\details{
For the generalized logistic g-function, the variance-covariance matrix of model parameters will be
of dimension (\code{len_beta} +3)x(\code{len_beta} +3), where \code{len_beta}  is the number of
beta coefficients in the model.
}
\seealso{
\code{\link{ocm}}
}

